FURNACE-112: split AddonRepository into two: AddonStateRepository & AddonStorageRepository to allow pluggable state (enabled/disabled) storage implementations without losing FileSystem backend for the storage
